--- src/pages/MainPage.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import '../styles/MainPage.css';
-
-const MainPage = () => {
-  return (
-    <div className="main-page">
-      <h1>메인 페이지에 오신 것을 환영합니다!</h1>
-      {/* 여기에 메인 페이지 컨텐츠를 추가하세요 */}
-    </div>
-  );
-};
-
-export default MainPage;
