+++ src/routes/AppRouter.jsx
@@ -1,14 +1,15 @@
-import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Example from "../pages/example/Example";
+import React from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import LandingPage from '../pages/landingpage/LandingPage';
 
-const AppRouter = () => (
-  <BrowserRouter>
-    <Routes>
-      <Route path="/" element={<Example />} /> // 여기 경로 설정으로 라우팅 설정
-      {/* 다른 라우트도 여기에 추가 */}
-    </Routes>
-  </BrowserRouter>
-);
+const AppRouter = () => {
+  return (
+    <Router>
+      <Routes>
+        <Route path="/" element={<LandingPage />} />
+      </Routes>
+    </Router>
+  );
+};
 
 export default AppRouter;
