--- src/styles/MainPage.css
@@ -1,17 +1,0 @@
-.main-page {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  background: linear-gradient(135deg, #45D3FF 0%, #CF83FF 100%);
-  color: white;
-  text-align: center;
-  padding: 20px;
-}
-
-.main-page h1 {
-  font-size: 2.5rem;
-  margin-bottom: 2rem;
-  text-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
-}
